alpha2-main/backend/shinhanalpha/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
executing==1.2.0
Flask==2.2.2
Flask-PyMongo==2.3.0
idna==3.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.36
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
tzdata==2022.7
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
